migrated qualifiers validations
