+++ services/ui-src/src/measures/2023/shared/Qualifiers/validationFunctions.ts
@@ -1,4 +1,12 @@
-import { ACS, ACSC, ACSM, CCS, CCSC, CCSM, HHCS } from "./validations";
+import {
+  ACS,
+  ACSC,
+  ACSM,
+  CCS,
+  CCSC,
+  CCSM,
+  HHCS,
+} from "shared/Qualifiers/validations";
 
 export const validationFunctions = {
   ACS,
