--- services/ui-src/src/measures/2023/shared/Qualifiers/validations/index.ts
@@ -1,5 +1,0 @@
-export * from "./adult";
-export * from "./childCombined";
-export * from "./childCHIP";
-export * from "./childMedicaid";
-export * from "./healthHome";
