+++ services/ui-src/src/measures/2024/shared/Qualifiers/validationFunctions.ts
@@ -1,4 +1,12 @@
-import { ACS, ACSM, ACSC, CCS, CCSC, CCSM, HHCS } from "./validations";
+import {
+  ACS,
+  ACSM,
+  ACSC,
+  CCS,
+  CCSC,
+  CCSM,
+  HHCS,
+} from "shared/Qualifiers/validations";
 
 export const validationFunctions = {
   ACS,
