+++ services/ui-src/src/shared/Qualifiers/validations/adultChip.ts
@@ -1,7 +1,7 @@
-import { ACSQualifierForm } from "../types";
-import { DeliverySystem } from "../types";
+import { ACSCQualifierForm } from "../../types/TypeQualifierForm";
+import { DeliverySystem } from "../../types/TypeQualifierForm";
 
-const validate21To64EqualsToOneHundredPercent = (data: ACSQualifierForm) => {
+const validate21To64EqualsToOneHundredPercent = (data: ACSCQualifierForm) => {
   const values = data["PercentageEnrolledInEachDeliverySystem"];
   const errorArray: any[] = [];
   const total21To64Percent = values?.reduce(
@@ -45,6 +45,4 @@ const validate21To64EqualsToOneHundredPercent = (data: ACSQualifierForm) => {
   return errorArray.length ? errorArray : [];
 };
 
-export const ACS = [validate21To64EqualsToOneHundredPercent];
-export const ACSM = [validate21To64EqualsToOneHundredPercent];
 export const ACSC = [validate21To64EqualsToOneHundredPercent];
